refactor: rename server and overview card files
