--- apps/dashboard/app/components/cards/overview-card/index.tsx
@@ -1,5 +1,0 @@
-import OverviewCardGrid from "./Grid";
-import type { OverviewCardProps } from "./Card";
-
-export default OverviewCardGrid;
-export type { OverviewCardProps };
